feat: add class DataModule
